# Conflicts:
#	Assets/Prefabs/EnemyMove.prefab
#	Assets/Prefabs/Player.prefab

+++ Assets/Scripts/ILife.cs
@@ -1,6 +1,6 @@
 ﻿
 public interface ILife
 {
-    void TakeHit(int damage);
+    void TakeHit(float damage);
 
 }
